fix: Reloading Placeholders
