+++ src/main/kotlin/net/blockventuremc/BlockVenture.kt
@@ -4,7 +4,7 @@ import io.github.cdimascio.dotenv.dotenv
 import net.blockventuremc.cache.PlayerCache
 import net.blockventuremc.database.DatabaseManager
 import net.blockventuremc.modules.i18n.TranslationCache
-import net.blockventuremc.modules.placeholders.registerPlaceholders
+import net.blockventuremc.modules.placeholders.PlayerPlaceholderManager
 import net.blockventuremc.utils.RegisterManager.registerAll
 import org.bukkit.Bukkit
 import org.bukkit.plugin.java.JavaPlugin
@@ -37,7 +37,8 @@ class BlockVenture: JavaPlugin() {
         TranslationCache.loadAll()
 
         logger.info("Registering placeholders...")
-        registerPlaceholders()
+        PlayerPlaceholderManager()
+
 
         logger.info("Registering modules...")
         registerAll()
